Added profanity check to automod

+++ services/automod.py
@@ -1,8 +1,10 @@
 import re
 import discord
+from datetime import timedelta
 
 async def automod(client: discord.Client, config):
     link_regex = r"((http|https)\:\/\/)?[a-zA-Z0-9\.\/\?\:@\-_=#]+\.([a-zA-Z]){2,6}([a-zA-Z0-9\.\&\/\?\:@\-_=#])*"
+    profanity_regex = r"pussy|cum|nude|nudes|nudities|fuck me|dick|cock|boobs|vagina|orgy|orgi|horny|horni|fuck|fuck you"
     
     @client.event
     async def on_message(message: discord.Message):
@@ -14,5 +16,13 @@ async def automod(client: discord.Client, config):
         if staff_role in message.author.roles:
             return
 
-        if re.search(link_regex, message.content):
+        if re.search(link_regex, message.content) and config["automod"]["filter_links"] == True:
+            await message.channel.send(f"<a:deny:1033973641030946826> {message.author.mention} Please refrain from sending links on the server!")
             await message.delete()
+            return
+
+        if re.search(profanity_regex, message.content, flags=re.IGNORECASE) and config["automod"]["filter_profanity"] == True:
+            await message.channel.send(f"<a:deny:1033973641030946826> {message.author.mention} Please refrain from using profanity on the server!")
+            await message.delete()
+            await message.author.timeout(timedelta(minutes=5))
+            return
